refactor(report): implemented requested changes from ratmir-rei-delas
