+++ bookmark/pom.xml
@@ -21,11 +21,6 @@
     <dependencies>
         <dependency>
             <groupId>pt.estga</groupId>
-            <artifactId>file</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>pt.estga</groupId>
             <artifactId>content</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
