+++ bookmark/src/main/java/pt/estga/bookmark/repositories/BookmarkRepository.java
@@ -2,7 +2,7 @@ package pt.estga.bookmark.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.estga.bookmark.entities.Bookmark;
-import pt.estga.file.enums.TargetType;
+import pt.estga.shared.enums.TargetType;
 
 import java.util.List;
 import java.util.Optional;
