+++ bookmark/src/main/java/pt/estga/bookmark/services/BookmarkService.java
@@ -1,7 +1,7 @@
 package pt.estga.bookmark.services;
 
 import pt.estga.bookmark.dtos.BookmarkDto;
-import pt.estga.file.enums.TargetType;
+import pt.estga.shared.enums.TargetType;
 import pt.estga.user.entities.User;
 
 import java.util.List;
