+++ report/src/main/java/pt/estga/report/dtos/ReportRequestDto.java
@@ -2,7 +2,7 @@ package pt.estga.report.dtos;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import pt.estga.file.enums.TargetType;
+import pt.estga.shared.enums.TargetType;
 import pt.estga.report.enums.ReportReason;
 
 public record ReportRequestDto(
