+++ report/src/main/java/pt/estga/report/repositories/ReportRepository.java
@@ -1,7 +1,7 @@
 package pt.estga.report.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import pt.estga.file.enums.TargetType;
+import pt.estga.shared.enums.TargetType;
 import pt.estga.report.entities.Report;
 
 import java.util.Optional;
