+++ shared/src/main/java/pt/estga/shared/enums/TargetType.java
@@ -1,4 +1,4 @@
-package pt.estga.file.enums;
+package pt.estga.shared.enums;
 
 public enum TargetType {
     MONUMENT, MARK, MARK_OCCURRENCE, GUILD, MASON,
